ElmyraService: Add initial implementation
This is a complete implementation of the Pixel Active Edge gesture
(codename Elmyra), which is a ContextHub-backed gesture present on the
Google Pixel 2, 2 XL, 3, 3 XL, and 3a phones. It uses hardware grip
sensors to detect when the user squeezes the sides of their phone (near
the bottom) and perform an action accordingly — which is opening the
Google Assistant on the stock ROM.

This service is almost fully modular in that it can just be built as a
package for applicable devices; no other integration is necessary, save
for some extra APIs in system_server. A Settings screen is also added
dynamically.

This improves upon the stock implementation of Active Edge by:
  - Exposing more sensitivity steps
  - Providing more actions to perform on trigger (currently: assistant,
    camera, flashlight, mute, power menu, screen on/off, and screenshot)
  - Vibrating with a "rejected" haptic effect if an action is
    temporarily unavailable
  - Adding a toggle for whether to enable the gesture when the screen is
    off

Note that this service requires the following commits in frameworks/base
to expose the necessary APIs for full functionality:
  - SystemUI: Allow privileged system apps to access screenshot service
  - core: Expose method to start assistant through Binder